+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 - `ItemAttribute` container struct for attributes.
 - `PartialOrd` and `Ord` traits to `StrangePartSet` and `SpellSet`.
 - `NONE` const for `SpellSet` and `StrangePartSet`.
-- `as_slice`, `as_slice_mut`, `iter_mut`, `iter_attributes`, `is_full`, `try_insert` methods to `SpellSet` and `StrangePartSet`.
+- `as_slice`, `as_slice_mut`, `iter_mut`, `iter_attributes`, `is_full`, `try_insert`, `insert_or_replace_last`, `replace`, methods to `SpellSet` and `StrangePartSet`.
 - `econ_defindex` method for `StockWeapon`.
 - `DescriptionFormat` and `EffectType` enums for traits `Attribute` and `Attributes`.
 - `Colored` trait for definitions associated with a color.